Moved welcome emails tests to a centralized location
